move: ErrorResponder 이동 ( root/audit/utils -> root/auth/utils )
